# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Có mặt trên tất cả nền tảng kỹ thuật số hàng đầu như Google Play, iOS, website. SandexCare là ứng dụng dịch vụ hỗ trợ khách hàng giải tỏa những khó khăn trong cuộc sống qua cuộc gọi 1-1 với chuyên viên lắng nghe của SandexCare, đem lại cảm hứng và động lực trong cuộc sống - công ty hàng đầu về sức khỏe tâm lý cộng đồng tại Việt Nam.\nSandexCare sẽ luôn bên bạn mọi lúc khi bạn cần nhất và hãy đến với chúng tôi để cùng nhau lan toả năng lượng tích cực cho cuộc sống để những vấn đề của chính bạn trở nên đơn giản hơn.\n\nFormat date: \"2022-01-04\".\nFormat time_slot: \"00:00-06:00\".\nFormat booking_time: \"01:00\".",
        "title": "Tài liệu đặc tả api.sandexcare.com",
        "termsOfService": "http://sandexcare.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/confirm-withdrawal": {
            "post": {
                "description": "Admin xác nhận yêu cầu rút tiền của chuyên viên/chuyên gia!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "API Confirm Withdrawal Request",
                "parameters": [
                    {
                        "description": "Input Confirm Withdrawal",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputConfirmWithdrawal"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/admin/users": {
            "get": {
                "description": "Admin xem tất cả người dùng trên hệ thống.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "API get all user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseGetAllUsers"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/admin/users/submit-diamond": {
            "post": {
                "description": "Admin cấp kim cương cho tài khoản user!",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "API Submit Diamonds User",
                "parameters": [
                    {
                        "description": "Input Submit Diamond",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InputSubmitDiamond"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/admin/withdrawal-history": {
            "get": {
                "description": "Admin xem tất cả yêu cầu rút tiền của cả hệ thống.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "API get all withdrawal history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseGetAllWithdrawalHistory"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/firebase/register": {
            "post": {
                "description": "Register Device Token - Đăng kí token device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Firebase"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token lấy từ API authen có prefix Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Token của người dùng lấy từ firebase, khi gọi lại API này, token cũ sẽ bị ghi đè (Xoá vĩnh viễn)",
                        "name": "Token",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.InputToken"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/firebase/send": {
            "post": {
                "description": "Send Background Notification By Firebase Service -  Gửi tin nhắn khi ngoại tuyến và trực tuyến",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Firebase"
                ],
                "parameters": [
                    {
                        "description": "Message: Tin nhắn gồm tựa đề, nội dung, hình ảnh và token của người gửi",
                        "name": "Message",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
        "/firebase/send/{uuid}": {
            "post": {
                "description": "Send Background Notification By Firebase Service BY UUID (TODO: Cần token service to service) -  Gửi tin nhắn khi ngoại tuyến và trực tuyến",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Firebase"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID của user",
                        "name": "uuid",
                        "in": "path"
                    },
                    {
                        "description": "Message: Tin nhắn gồm tựa đề, nội dung, hình ảnh và token của người gửi",
                        "name": "Message",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.GetAllUsersData": {
            "type": "object",
            "properties": {
                "diamond": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.GetAllWithdrawalHistoryData": {
            "type": "object",
            "properties": {
                "amount_money": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "listener_id": {
                    "type": "string"
                },
                "listener_name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "model.InputConfirmWithdrawal": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.InputSubmitDiamond": {
            "type": "object",
            "properties": {
                "diamond": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "model.InputToken": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "notification": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "model.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ResponseGetAllUsers": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GetAllUsersData"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
        "model.ResponseGetAllWithdrawalHistory": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.GetAllWithdrawalHistoryData"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Success"
                }
            }
        },
...
        "model.ResponseSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}